refactor: move player into subdirectory
