+++ components/Player/index.tsx
